+++ js/string-length.js
@@ -1,0 +1,7 @@
+// Функция максимальной длины строки
+
+function isStringValid(string, maxLength) {
+  return string.length <= maxLength;
+}
+
+isStringValid('Текст', 140);
